+++ src/users/client.js
@@ -1,18 +1,22 @@
 import axios from "axios";
+export const USERS_API = `http://localhost:4000/api/users`;
 
 const request = axios.create({
   withCredentials: true,
 });
 
-const USERS_API = "http://localhost:4000/api/users";
-
-export const findAllUsers = async () => {
-  const response = await axios.get(USERS_API);
+export const signin = async (credentials) => {
+  const response = await request.post(`${USERS_API}/signin`, credentials);
   return response.data;
 };
 
-export const signin = async (credentials) => {
-  const response = await request.post(`${USERS_API}/signin`, credentials);
+export const signout = async () => {
+  const response = await request.post(`${USERS_API}/signout`);
+  return response.data;
+};
+
+export const signup = async (credentials) => {
+  const response = await request.post(`${USERS_API}/signup`, credentials);
   return response.data;
 };
 
@@ -26,6 +30,10 @@ export const updateUser = async (user) => {
   return response.data;
 };
 
+export const findAllUsers = async () => {
+  const response = await request.get(`${USERS_API}`);
+  return response.data;
+};
 export const createUser = async (user) => {
   const response = await request.post(`${USERS_API}`, user);
   return response.data;
@@ -35,18 +43,7 @@ export const findUserById = async (id) => {
   const response = await request.get(`${USERS_API}/${id}`);
   return response.data;
 };
-
 export const deleteUser = async (user) => {
   const response = await request.delete(`${USERS_API}/${user._id}`);
-  return response.data;
-};
-
-export const signup = async (credentials) => {
-  const response = await request.post(`${USERS_API}/signup`, credentials);
-  return response.data;
-};
-
-export const signout = async () => {
-  const response = await request.post(`${USERS_API}/signout`);
   return response.data;
 };
